Modificado para no eliminar odontologo con paciente asignado

+++ IntegradorII (2)/IntegradorII/src/main/java/com/example/IntegradorII/service/OdontologoService.java
@@ -1,8 +1,10 @@
 package com.example.IntegradorII.service;
-
+import com.example.IntegradorII.entity.Paciente;
+import com.example.IntegradorII.exception.BadRequestException;
 import com.example.IntegradorII.entity.Domicilio;
 import com.example.IntegradorII.entity.Odontologo;
 import com.example.IntegradorII.repository.OdontologoRepository;
+import com.example.IntegradorII.repository.PacienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,13 +16,23 @@ public class OdontologoService {
     @Autowired
     private OdontologoRepository odontologoRepository;
 
+    @Autowired
+    private PacienteRepository pacienteRepository;
+
     public Odontologo guardarOdontologo(Odontologo odontologo){
         return odontologoRepository.save(odontologo);
     }
     public List<Odontologo> listarOdontologos(){
         return odontologoRepository.findAll();
     }
-    public void eliminarOdontologo(Long id){
+    public void eliminarOdontologo(Long id) throws BadRequestException {
+        // Verifica si el odontólogo tiene pacientes asignados
+        List<Paciente> pacientes = pacienteRepository.findByOdontologoId(id);
+        if (!pacientes.isEmpty()) {
+            throw new BadRequestException("No se puede eliminar el odontólogo, tiene pacientes asignados.");
+        }
+
+        // Si no tiene pacientes, el odontólogo puede ser eliminado
         odontologoRepository.deleteById(id);
     }
     public Optional<Odontologo> buscarPorID(Long id){ return odontologoRepository.findById(id);}
